Remove service from api
