+++ pkg/service.go
@@ -9,7 +9,6 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/blocknative/dreamboat/metrics"
 	"github.com/blocknative/dreamboat/pkg/structs"
 	"github.com/flashbots/go-boost-utils/types"
 	ds "github.com/ipfs/go-datastore"
@@ -24,38 +23,12 @@ var (
 	ErrBeaconNodeSyncing = errors.New("beacon node is syncing")
 )
 
-type Relay interface {
-	// Proposer API
-	RegisterValidator(context.Context, []structs.SignedValidatorRegistration) error
-	GetHeader(context.Context, structs.HeaderRequest) (*types.GetHeaderResponse, error)
-	GetPayload(context.Context, *types.SignedBlindedBeaconBlock) (*types.GetPayloadResponse, error)
-
-	// Builder APIs
-	SubmitBlock(context.Context, *types.BuilderSubmitBlockRequest) error
-	GetValidators() structs.BuilderGetValidatorsResponseEntrySlice
-
-	AttachMetrics(m *metrics.Metrics)
-}
-
 type Datastore interface {
-	GetHeadersBySlot(ctx context.Context, slot uint64) ([]structs.HeaderAndTrace, error)
-	GetHeadersByBlockHash(ctx context.Context, hash types.Hash) ([]structs.HeaderAndTrace, error)
-	GetHeadersByBlockNum(ctx context.Context, num uint64) ([]structs.HeaderAndTrace, error)
-	GetLatestHeaders(ctx context.Context, limit uint64, stopLag uint64) ([]structs.HeaderAndTrace, error)
-
-	PutDelivered(context.Context, structs.Slot, structs.DeliveredTrace, time.Duration) error
-	GetDelivered(context.Context, structs.PayloadQuery) (structs.BidTraceWithTimestamp, error)
-	GetDeliveredBatch(context.Context, []structs.PayloadQuery) ([]structs.BidTraceWithTimestamp, error)
-	PutPayload(context.Context, structs.PayloadKey, *structs.BlockBidAndTrace, time.Duration) error
-	GetPayload(context.Context, structs.PayloadKey) (*structs.BlockBidAndTrace, bool, error)
-	PutRegistrationRaw(context.Context, structs.PubKey, []byte, time.Duration) error
 	GetRegistration(context.Context, structs.PubKey) (types.SignedValidatorRegistration, error)
 }
-
 type Service struct {
 	Log             log.Logger
 	Config          Config
-	Relay           Relay
 	Datastore       Datastore
 	NewBeaconClient func() (BeaconClient, error)
 
@@ -68,12 +41,11 @@ type Service struct {
 	updateTime   atomic.Value
 }
 
-func NewService(l log.Logger, c Config, d Datastore, r Relay, as *AtomicState) *Service {
+func NewService(l log.Logger, c Config, d Datastore, as *AtomicState) *Service {
 	return &Service{
 		Log:       l.WithField("relay-service", "Service"),
 		Config:    c,
 		Datastore: d,
-		Relay:     r,
 		state:     as,
 	}
 }
@@ -260,115 +232,4 @@ func (s *Service) updateKnownValidators(ctx context.Context, client BeaconClient
 	s.state.validators.Store(state)
 
 	return nil
-}
-
-func (s *Service) RegisterValidator(ctx context.Context, payload []structs.SignedValidatorRegistration) error {
-	return s.Relay.RegisterValidator(ctx, payload)
-}
-
-func (s *Service) GetHeader(ctx context.Context, request structs.HeaderRequest) (*types.GetHeaderResponse, error) {
-	return s.Relay.GetHeader(ctx, request)
-}
-
-func (s *Service) GetPayload(ctx context.Context, payloadRequest *types.SignedBlindedBeaconBlock) (*types.GetPayloadResponse, error) {
-	return s.Relay.GetPayload(ctx, payloadRequest)
-}
-
-func (s *Service) SubmitBlock(ctx context.Context, submitBlockRequest *types.BuilderSubmitBlockRequest) error {
-	return s.Relay.SubmitBlock(ctx, submitBlockRequest)
-}
-
-func (s *Service) GetValidators() structs.BuilderGetValidatorsResponseEntrySlice {
-	return s.Relay.GetValidators()
-}
-
-func (s *Service) Registration(ctx context.Context, pk types.PublicKey) (types.SignedValidatorRegistration, error) {
-	return s.Datastore.GetRegistration(ctx, structs.PubKey{PublicKey: pk})
-}
-
-func (s *Service) GetPayloadDelivered(ctx context.Context, query structs.PayloadTraceQuery) ([]structs.BidTraceExtended, error) {
-	var (
-		event structs.BidTraceWithTimestamp
-		err   error
-	)
-
-	if query.HasSlot() {
-		event, err = s.Datastore.GetDelivered(ctx, structs.PayloadQuery{Slot: query.Slot})
-	} else if query.HasBlockHash() {
-		event, err = s.Datastore.GetDelivered(ctx, structs.PayloadQuery{BlockHash: query.BlockHash})
-	} else if query.HasBlockNum() {
-		event, err = s.Datastore.GetDelivered(ctx, structs.PayloadQuery{BlockNum: query.BlockNum})
-	} else if query.HasPubkey() {
-		event, err = s.Datastore.GetDelivered(ctx, structs.PayloadQuery{PubKey: query.Pubkey})
-	} else {
-		return s.getTailDelivered(ctx, query.Limit, query.Cursor)
-	}
-
-	if err == nil {
-		return []structs.BidTraceExtended{{BidTrace: event.BidTrace, BlockNumber: event.BlockNumber, NumTx: event.NumTx}}, err
-	} else if errors.Is(err, ds.ErrNotFound) {
-		return []structs.BidTraceExtended{}, nil
-	}
-	return nil, err
-}
-
-func (s *Service) getTailDelivered(ctx context.Context, limit, cursor uint64) ([]structs.BidTraceExtended, error) {
-	headSlot := s.state.Beacon().HeadSlot()
-	start := headSlot
-	if cursor != 0 {
-		start = min(headSlot, structs.Slot(cursor))
-	}
-
-	stop := start - structs.Slot(s.Config.TTL/DurationPerSlot)
-
-	batch := make([]structs.BidTraceWithTimestamp, 0, limit)
-	queries := make([]structs.PayloadQuery, 0, limit)
-
-	for highSlot := start; len(batch) < int(limit) && stop <= highSlot; highSlot -= structs.Slot(limit) {
-		queries = queries[:0]
-		for s := highSlot; highSlot-structs.Slot(limit) < s && stop <= s; s-- {
-			queries = append(queries, structs.PayloadQuery{Slot: s})
-		}
-
-		nextBatch, err := s.Datastore.GetDeliveredBatch(ctx, queries)
-		if err != nil {
-			s.Log.WithError(err).Warn("failed getting header batch")
-		} else {
-			batch = append(batch, nextBatch[:min(int(limit)-len(batch), len(nextBatch))]...)
-		}
-	}
-
-	events := make([]structs.BidTraceExtended, 0, len(batch))
-	for _, event := range batch {
-		events = append(events, event.BidTraceExtended)
-	}
-	return events, nil
-}
-
-func (s *Service) GetBlockReceived(ctx context.Context, query structs.HeaderTraceQuery) ([]structs.BidTraceWithTimestamp, error) {
-	var (
-		events []structs.HeaderAndTrace
-		err    error
-	)
-
-	if query.HasSlot() {
-		events, err = s.Datastore.GetHeadersBySlot(ctx, uint64(query.Slot))
-	} else if query.HasBlockHash() {
-		events, err = s.Datastore.GetHeadersByBlockHash(ctx, query.BlockHash)
-	} else if query.HasBlockNum() {
-		events, err = s.Datastore.GetHeadersByBlockNum(ctx, query.BlockNum)
-	} else {
-		events, err = s.Datastore.GetLatestHeaders(ctx, query.Limit, uint64(s.Config.TTL/DurationPerSlot))
-	}
-
-	if err == nil {
-		traces := make([]structs.BidTraceWithTimestamp, 0, len(events))
-		for _, event := range events {
-			traces = append(traces, *event.Trace)
-		}
-		return traces, err
-	} else if errors.Is(err, ds.ErrNotFound) {
-		return []structs.BidTraceWithTimestamp{}, nil
-	}
-	return nil, err
 }
